Add test kitchen for build nodes
We want to create build nodes that can run delivery jobs from the
local workstation instead of having to go through the Chef Delivery
pipeline to do the verify phase steps. We want this so we're ensuring
verification runs on the same platforms as the build nodes that would
be in delivery.